testsolving day 1

+++ algorithms/countbits.cpp
@@ -1,5 +1,5 @@
-int countBits(int i) {
-    int count = 0;
+ll countBits(ll i) {
+    ll count = 0;
     while (i != 0) {
         count += (i & 1);
         i >>= 1;
